requirements.txt
asttokens==2.4.1
colorama==0.4.6
comm==0.2.0
covabra-mail @ file://srvcic/b/Bibliotecas/covabra-mail-3.0.0.tar.gz#sha256=20c057a2ad51c64b23cf7b68ec1ddc17b29d7a876fa6fcb5f8e24773d786fb8e
debugpy==1.8.0
decorator==5.1.1
executing==2.0.1
greenlet==3.0.2
ipykernel==6.27.1
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.2
jupyter_client==8.6.0
jupyter_core==5.5.1
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
numpy==1.26.2
packaging==23.2
pandas==2.1.4
parso==0.8.3
platformdirs==4.1.0
prompt-toolkit==3.0.43
psutil==5.9.7
pure-eval==0.2.2
Pygments==2.17.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pywin32==306
pyzmq==25.1.2
secure-smtplib==0.1.1
six==1.16.0
SQLAlchemy==2.0.23
stack-data==0.6.3
tornado==6.4
traitlets==5.14.0
typing_extensions==4.9.0
tzdata==2023.3
wcwidth==0.2.12
